Use actual generated Java code from WebGME for 'GenX' examples. Only packages/imports corrected for local project package structure.

+++ src/java/ucef-java-core/src/main/java/gov/nist/ucef/hla/base/HLACodecUtils.java
@@ -1,17 +1,17 @@
 /*
- * This software is contributed as a public service by The National Institute of Standards 
+ * This software is contributed as a public service by The National Institute of Standards
  * and Technology (NIST) and is not subject to U.S. Copyright
- * 
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this 
- * software and associated documentation files (the "Software"), to deal in the Software 
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this
+ * software and associated documentation files (the "Software"), to deal in the Software
  * without restriction, including without limitation the rights to use, copy, modify,
  * merge, publish, distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to the following 
+ * permit persons to whom the Software is furnished to do so, subject to the following
  * conditions:
- * 
+ *
  * The above NIST contribution notice and this permission and disclaimer notice shall be
  * included in all copies or substantial portions of the Software.
- * 
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
  * INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
  * PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
@@ -57,15 +57,15 @@ public class HLACodecUtils
 			throw new UCEFException("Failed to initialize HLA encode/decode utilities.", e);
 		}
 	}
-	
+
 	/**
 	 * Decode HLA byte array representation of a bye
-	 * 
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param bytes the HLA byte array
 	 * @return the decoded value
 	 */
-	public static short asByte( EncoderFactory encoderFactory, byte[] bytes )
+	public static byte asByte( EncoderFactory encoderFactory, byte[] bytes )
 	{
 		HLAbyte value = makeHLAType(encoderFactory, (byte)0);
 		try
@@ -80,10 +80,10 @@ public class HLACodecUtils
 			                         bytes.length );
 		}
 	}
-	
+
 	/**
 	 * Decode HLA byte array representation of a short
-	 * 
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param bytes the HLA byte array
 	 * @return the decoded value
@@ -103,17 +103,17 @@ public class HLACodecUtils
 			                         bytes.length );
 		}
 	}
-	
+
 	/**
 	 * Decode HLA byte array representation of an integer
-	 * 
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param bytes the HLA byte array
 	 * @return the decoded value
 	 */
 	public static int asInt( EncoderFactory encoderFactory, byte[] bytes )
 	{
-		HLAinteger32BE value = makeHLAType(encoderFactory, (int)0);
+		HLAinteger32BE value = makeHLAType(encoderFactory, 0);
 		try
 		{
 			value.decode( bytes );
@@ -126,10 +126,10 @@ public class HLACodecUtils
 			                         bytes.length );
 		}
 	}
-	
+
 	/**
 	 * Decode HLA byte array representation of a long
-	 * 
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param bytes the HLA byte array
 	 * @return the decoded value
@@ -149,10 +149,10 @@ public class HLACodecUtils
 			                         bytes.length );
 		}
 	}
-	
+
 	/**
 	 * Decode HLA byte array representation of a float
-	 * 
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param bytes the HLA byte array
 	 * @return the decoded value
@@ -172,10 +172,10 @@ public class HLACodecUtils
 			                         bytes.length );
 		}
 	}
-	
+
 	/**
 	 * Decode HLA byte array representation of a double
-	 * 
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param bytes the HLA byte array
 	 * @return the decoded value
@@ -195,10 +195,10 @@ public class HLACodecUtils
 			                         bytes.length );
 		}
 	}
-	
+
 	/**
 	 * Decode HLA byte array representation of a boolean
-	 * 
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param bytes the HLA byte array
 	 * @return the decoded value
@@ -221,7 +221,7 @@ public class HLACodecUtils
 
 	/**
 	 * Decode HLA byte array representation of a char
-	 * 
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param bytes the HLA byte array
 	 * @return the decoded value
@@ -241,10 +241,10 @@ public class HLACodecUtils
 			                         bytes.length );
 		}
 	}
-	
+
 	/**
 	 * Decode HLA byte array representation of a unicode string
-	 * 
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param bytes the HLA byte array
 	 * @return the decoded value
@@ -256,12 +256,12 @@ public class HLACodecUtils
 		{
 			value.decode( bytes );
 			String result = value.getValue();
-			
+
 			// check for null terminator at end of string and remove as necessary
 			int length = result.length();
 			if( length > 0 && result.charAt( length - 1 ) == 0 )
 				result = result.substring( 0, length - 1  );
-			
+
 			return result;
 		}
 		catch( DecoderException de )
@@ -273,92 +273,92 @@ public class HLACodecUtils
 	}
 
 	/**
-	 * Encode a short to an HLA byte array representation 
-	 * 
+	 * Encode a short to an HLA byte array representation
+	 *
 	 * @param encoderFactory the encoder instance to use
-	 * @param value the value to encode 
+	 * @param value the value to encode
 	 * @return the HLA byte array
 	 */
 	public static byte[] encode(EncoderFactory encoderFactory, char value)
 	{
 		return makeHLAType(encoderFactory, value).toByteArray();
 	}
-	
+
 	/**
-	 * Encode a byte to an HLA byte array representation 
-	 * 
+	 * Encode a byte to an HLA byte array representation
+	 *
 	 * @param encoderFactory the encoder instance to use
-	 * @param value the value to encode 
+	 * @param value the value to encode
 	 * @return the HLA byte array
 	 */
 	public static byte[] encode(EncoderFactory encoderFactory, byte value)
 	{
 		return makeHLAType(encoderFactory, value).toByteArray();
 	}
-	
+
 	/**
-	 * Encode a short to an HLA byte array representation 
-	 * 
+	 * Encode a short to an HLA byte array representation
+	 *
 	 * @param encoderFactory the encoder instance to use
-	 * @param value the value to encode 
+	 * @param value the value to encode
 	 * @return the HLA byte array
 	 */
 	public static byte[] encode(EncoderFactory encoderFactory, short value)
 	{
 		return makeHLAType(encoderFactory, value).toByteArray();
 	}
-	
+
 	/**
-	 * Encode an integer to an HLA byte array representation 
-	 * 
+	 * Encode an integer to an HLA byte array representation
+	 *
 	 * @param encoderFactory the encoder instance to use
-	 * @param value the value to encode 
+	 * @param value the value to encode
 	 * @return the HLA byte array
 	 */
 	public static byte[] encode(EncoderFactory encoderFactory, int value)
 	{
 		return makeHLAType(encoderFactory, value).toByteArray();
 	}
-	
+
 	/**
-	 * Encode a long to an HLA byte array representation 
-	 * 
+	 * Encode a long to an HLA byte array representation
+	 *
 	 * @param encoderFactory the encoder instance to use
-	 * @param value the value to encode 
+	 * @param value the value to encode
 	 * @return the HLA byte array
 	 */
 	public static byte[] encode(EncoderFactory encoderFactory, long value)
 	{
 		return makeHLAType(encoderFactory, value).toByteArray();
 	}
-	
+
 	/**
-	 * Encode a double to an HLA byte array representation 
-	 * 
+	 * Encode a double to an HLA byte array representation
+	 *
 	 * @param encoderFactory the encoder instance to use
-	 * @param value the value to encode 
+	 * @param value the value to encode
 	 * @return the HLA byte array
 	 */
 	public static byte[] encode(EncoderFactory encoderFactory, double value)
 	{
 		return makeHLAType(encoderFactory, value).toByteArray();
 	}
-	
+
 	/**
-	 * Encode a float to an HLA byte array representation 
-	 * 
+	 * Encode a float to an HLA byte array representation
+	 *
 	 * @param encoderFactory the encoder instance to use
-	 * @param value the value to encode 
+	 * @param value the value to encode
 	 * @return the HLA byte array
 	 */
 	public static byte[] encode(EncoderFactory encoderFactory, float value)
 	{
 		return makeHLAType(encoderFactory, value).toByteArray();
 	}
-	
+
 	/**
-	 * Encode a boolean to an HLA byte array representation 
-	 * 
+	 * Encode a boolean to an HLA byte array representation
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param value the value to encode
 	 * @return the HLA byte array
@@ -367,10 +367,10 @@ public class HLACodecUtils
 	{
 		return makeHLAType(encoderFactory, value).toByteArray();
 	}
-	
+
 	/**
 	 * Encode a string to an HLA byte array representation of a unicode string
-	 * 
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param value the value to encode
 	 * @return the HLA byte array
@@ -379,10 +379,10 @@ public class HLACodecUtils
 	{
 		return makeHLAType( encoderFactory, value ).toByteArray();
 	}
-	
+
 	/**
-	 * Encode a byte to an {@link HLAbyte} representation 
-	 * 
+	 * Encode a byte to an {@link HLAbyte} representation
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param value the value to encode
 	 * @return the {@link HLAbyte}
@@ -391,10 +391,10 @@ public class HLACodecUtils
 	{
 		return encoderFactory.createHLAbyte(value);
 	}
-	
+
 	/**
-	 * Encode a short to an {@link HLAinteger16BE} representation 
-	 * 
+	 * Encode a short to an {@link HLAinteger16BE} representation
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param value the value to encode
 	 * @return the {@link HLAinteger16BE}
@@ -403,10 +403,10 @@ public class HLACodecUtils
 	{
 		return encoderFactory.createHLAinteger16BE(value);
 	}
-	
+
 	/**
-	 * Encode an integer to an {@link HLAinteger32BE} representation 
-	 * 
+	 * Encode an integer to an {@link HLAinteger32BE} representation
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param value the value to encode
 	 * @return the {@link HLAinteger32BE}
@@ -415,10 +415,10 @@ public class HLACodecUtils
 	{
 		return encoderFactory.createHLAinteger32BE(value);
 	}
-	
+
 	/**
-	 * Encode a long to an {@link HLAinteger64BE} representation 
-	 * 
+	 * Encode a long to an {@link HLAinteger64BE} representation
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param value the value to encode
 	 * @return the HLA byte array
@@ -427,10 +427,10 @@ public class HLACodecUtils
 	{
 		return encoderFactory.createHLAinteger64BE(value);
 	}
-	
+
 	/**
-	 * Encode a double to an {@link HLAfloat64BE} representation 
-	 * 
+	 * Encode a double to an {@link HLAfloat64BE} representation
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param value the value to encode
 	 * @return the {@link HLAfloat64BE}
@@ -439,10 +439,10 @@ public class HLACodecUtils
 	{
 		return encoderFactory.createHLAfloat64BE(value);
 	}
-	
+
 	/**
-	 * Encode a float to an {@link HLAfloat32BE} representation 
-	 * 
+	 * Encode a float to an {@link HLAfloat32BE} representation
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param value the value to encode
 	 * @return the {@link HLAfloat32BE}
@@ -451,10 +451,10 @@ public class HLACodecUtils
 	{
 		return encoderFactory.createHLAfloat32BE(value);
 	}
-	
+
 	/**
-	 * Encode a boolean to an {@link HLAboolean} representation 
-	 * 
+	 * Encode a boolean to an {@link HLAboolean} representation
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param value the value to encode
 	 * @return the {@link HLAboolean}
@@ -463,10 +463,10 @@ public class HLACodecUtils
 	{
 		return encoderFactory.createHLAboolean(value);
 	}
-	
+
 	/**
-	 * Encode a char to an {@link HLAunicodeChar} representation 
-	 * 
+	 * Encode a char to an {@link HLAunicodeChar} representation
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param value the value to encode
 	 * @return the {@link HLAunicodeChar}
@@ -474,13 +474,13 @@ public class HLACodecUtils
 	private static HLAunicodeChar makeHLAType(EncoderFactory encoderFactory, char value)
 	{
 		// NOTE: we need to cast the char to a short here
-		//       in order to create the HLAunicodeChar 
+		//       in order to create the HLAunicodeChar
 		return encoderFactory.createHLAunicodeChar((short)value);
 	}
-	
+
 	/**
-	 * Encode a string to an {@link HLAunicodeString} representation 
-	 * 
+	 * Encode a string to an {@link HLAunicodeString} representation
+	 *
 	 * @param encoderFactory the encoder instance to use
 	 * @param value the value to encode
 	 * @return the {@link HLAunicodeString}
